+++ stories/treemap/9_zero_values.tsx
@@ -16,75 +16,43 @@
  * specific language governing permissions and limitations
  * under the License. */
 
-import { Chart, Datum, Partition, PartitionLayout, Settings } from '../../src';
+import { Chart, Datum, Partition, PartitionLayout } from '../../src';
 import { mocks } from '../../src/mocks/hierarchical/index';
 import { config } from '../../src/chart_types/partition_chart/layout/config/config';
 import { arrayToLookup, hueInterpolator } from '../../src/chart_types/partition_chart/layout/utils/calcs';
-import { countryDimension, productDimension } from '../../src/mocks/hierarchical/dimension_codes';
-
+import { productDimension } from '../../src/mocks/hierarchical/dimension_codes';
 import { palettes } from '../../src/mocks/hierarchical/palettes';
 import React from 'react';
-import { ShapeTreeNode } from '../../src/chart_types/partition_chart/layout/types/viewmodel_types';
 
 const productLookup = arrayToLookup((d: Datum) => d.sitc1, productDimension);
-const countryLookup = arrayToLookup((d: Datum) => d.country, countryDimension);
 
-const interpolatorTurbo = hueInterpolator(palettes.turbo.map(([r, g, b]) => [r, g, b, 0.7]));
+// style calcs
+const interpolatorCET2s = hueInterpolator(palettes.CET2s.map(([r, g, b]) => [r, g, b, 0.7]));
+
+const defaultFillColor = (colorMaker: any) => (d: any, i: number, a: any[]) => colorMaker(i / (a.length + 1));
 
 export const example = () => (
   <Chart className="story-chart">
-    <Settings showLegend />
     <Partition
       id="spec_1"
-      data={mocks.sunburst}
+      data={mocks.pie.map((d: any, i: number) => (i ? d : { ...d, exportVal: 0 }))}
       valueAccessor={(d: Datum) => d.exportVal as number}
       valueFormatter={(d: number) => `$${config.fillLabel.valueFormatter(Math.round(d / 1000000000))}\xa0Bn`}
       layers={[
         {
           groupByRollup: (d: Datum) => d.sitc1,
-          nodeLabel: (d: any) => productLookup[d].name.toUpperCase(),
+          nodeLabel: (d: Datum) => productLookup[d].name,
           fillLabel: {
-            valueFormatter: () => '',
-            fontFamily: 'Helvetica',
-            textColor: 'grey',
             textInvertible: true,
-          },
-          shape: {
-            fillColor: 'rgba(0, 0, 0, 0)',
-          },
-        },
-        {
-          groupByRollup: (d: Datum) => d.dest,
-          nodeLabel: (d: any) => countryLookup[d].name,
-          fillLabel: {
             valueFormatter: (d: number) => `${config.fillLabel.valueFormatter(Math.round(d / 1000000000))}\xa0Bn`,
-            textColor: 'black',
-            textInvertible: true,
-            fontWeight: 100,
-            fontStyle: 'normal',
-            fontFamily: 'Helvetica',
-            fontVariant: 'normal',
-            valueFont: {
-              fontWeight: 100,
-            },
           },
           shape: {
-            fillColor: (d: ShapeTreeNode) => {
-              // primarily, pick color based on parent's index, but then perturb by the index within the parent
-              return interpolatorTurbo(
-                (d.parent.sortIndex + d.sortIndex / d.parent.children.length) / (d.parent.parent.children.length + 1),
-              );
-            },
+            fillColor: defaultFillColor(interpolatorCET2s),
           },
         },
       ]}
       config={{
         partitionLayout: PartitionLayout.treemap,
-        margin: { top: 0, bottom: 0, left: 0, right: 0 },
-        minFontSize: 4,
-        maxFontSize: 84,
-        idealFontSizeJump: 1.05,
-        outerSizeRatio: 1,
       }}
     />
   </Chart>